fixing test cases, adding auth, implimented jsr303 on console

+++ gamestore/src/main/java/com/company/gamestore/repositories/GameRepo.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface GameRepo extends JpaRepository<Game, Long> {
+public interface GameRepo extends JpaRepository<Game, Integer> {
     List<Game> findByTitle(String title);
 }
